woo middleware auth working
