+++ server/token.go
@@ -12,98 +12,6 @@ import (
 	"github.com/google/uuid"
 )
 
-// AuthKey is a key in the context of a request populated by an AuthHandler.
-type AuthKey int64
-
-const (
-	AuthLoggedIn AuthKey = iota
-	AuthUser
-)
-
-// AuthHandler is middleware that will accept a request, extract the token used
-// for authentication, and make calls to get a User entity that represents the
-// logged in user from the token.
-//
-// Keys are added to the request context before the request is passed to the
-// next step in the chain. AuthUser will contain the logged-in user, and
-// AuthLoggedIn will return whether the user is logged in (only applies for
-// optional logins; for non-optional, not being logged in will result in an
-// HTTP error being returned before the request is passed to the next handler).
-type AuthHandler struct {
-	db            dao.UserRepository
-	secret        []byte
-	required      bool
-	defaultUser   dao.User
-	unauthedDelay time.Duration
-	next          http.Handler
-}
-
-func (ah *AuthHandler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
-	var loggedIn bool
-	user := ah.defaultUser
-
-	tok, err := getJWT(req)
-	if err != nil {
-		// deliberately leaving as embedded if instead of &&
-		if ah.required {
-			// error here means token isn't present (or at least isn't in the
-			// expected format, which for all intents and purposes is non-existent).
-			// This is not okay if auth is required.
-
-			result := jsonUnauthorized("", err.Error())
-			time.Sleep(ah.unauthedDelay)
-			result.writeResponse(w, req)
-			return
-		}
-	} else {
-		// validate the token
-		lookupUser, err := validateAndLookupJWTUser(req.Context(), tok, ah.secret, ah.db)
-		if err != nil {
-			// deliberately leaving as embedded if instead of &&
-			if ah.required {
-				// there was a validation error. the user does not count as logged in.
-				// if logging in is required, that's not okay.
-
-				result := jsonUnauthorized("", err.Error())
-				time.Sleep(ah.unauthedDelay)
-				result.writeResponse(w, req)
-				return
-			}
-		} else {
-			user = lookupUser
-			loggedIn = true
-		}
-	}
-
-	ctx := req.Context()
-	ctx = context.WithValue(ctx, AuthLoggedIn, loggedIn)
-	ctx = context.WithValue(ctx, AuthUser, user)
-	req = req.WithContext(ctx)
-	ah.next.ServeHTTP(w, req)
-}
-
-func RequireAuth(db dao.UserRepository, secret []byte, unauthDelay time.Duration, defaultUser dao.User, next http.Handler) *AuthHandler {
-	return &AuthHandler{
-		db:            db,
-		secret:        secret,
-		unauthedDelay: unauthDelay,
-		defaultUser:   dao.User{},
-		required:      true,
-		next:          next,
-	}
-}
-
-func OptionalAuth(db dao.UserRepository, secret []byte, unauthDelay time.Duration, defaultUser dao.User, next http.Handler) *AuthHandler {
-	return &AuthHandler{
-		db:            db,
-		secret:        secret,
-		unauthedDelay: unauthDelay,
-		defaultUser:   dao.User{},
-		required:      false,
-		next:          next,
-	}
-}
-
 func validateAndLookupJWTUser(ctx context.Context, tok string, secret []byte, db dao.UserRepository) (dao.User, error) {
 	var user dao.User
